refactor: move Advocate interface to `/types/`
