+++ hooks/advocate-conversion-utils.ts
@@ -2,7 +2,6 @@ import Airtable from 'airtable'
 
 // TODO: Understand why this import works with '../' and not with '~/'
 import {
-  Advocate,
   AdvocatesWorldRegion
 } from '../store/modules/advocates'
 
@@ -10,6 +9,7 @@ import {
   getImageUrl,
   findImageAttachment
 } from './airtable-conversion-utils'
+import { Advocate } from '~/types/advocate'
 
 const RECORD_FIELDS = Object.freeze({
   name: 'Name',
