--- types/advocate.d.ts
@@ -1,6 +1,0 @@
-export interface Advocate {
-  image: string,
-  location: string,
-  name: string,
-  region: string
-}
